Add gzip and extension options for dump output
The code has been updated to allow users to specify whether gzip compression should be used for dump output. An option to specify an extension for the dump file has also been added. Furthermore, a test for the function getExportName has been implemented to ensure it correctly generates the expected export filenames